Separate components + remove proxy
